separated CustomTimePickerDialog
